Move SimpleCar files into Cars, and remove SimpleCar directory.
