+++ drake/examples/Cars/lcm_vector_gen.py
