+++ drake/examples/Cars/models/boxcar.urdf
