+++ drake/examples/Cars/simple_car.cc
