+++ drake/examples/Cars/simple_car.h
