+++ drake/examples/Cars/simple_car_demo.sh
