+++ drake/examples/Cars/simple_car_gen.sh
