+++ drake/examples/Cars/test/simple_car_test.cc
